1. Update location module package structure 2. Minor refactor
