+++ syrf-location/src/main/java/com/syrf/location/utils/Constants.kt
@@ -1,4 +1,4 @@
-package utils
+package com.syrf.location.utils
 
 object Constants {
