+++ syrf-location/src/main/java/com/syrf/location/utils/Types.kt
@@ -1,4 +1,4 @@
-package utils
+package com.syrf.location.utils
 
 import android.location.Location
